The power-spectra of the normalized waveforms (used for cross-correlations) is now un-normalized. I have confirmed that this agrees with Parseval's Theorem (approximately, as I am dropping the Nyquist frequency bin). The beginnings of simple filtering is now done in the summing power spectra kernel, so far this is as simple as zeroing things outside of the pass band. More filter logic is to be added here.
There is really no need to output the normalized waveform once I have the RMS, I could simply fourier transform the un-normalized version and then divide each bin by the square of the RMS of the time domain waveform.
However, this can be changed later if it is important.
This is a commit where everything works.